Tighter cleanup work for verifier (#106)
Previously this either cleaned up all builds if enabled, or left detritus everywhere if disabled. This PR changes the disabled behaviour so that it cleans up successful builds as there's no user journey that involves investigating that. When a failed build is left about, it now prints out a more identifiable message in the log that makes it easier for an investigator to find the directory in order to get the built artifacts that don't match.