Converted build verifier into a cobra tool (#112)
The current behaviour of running continuously has been turned into a command called continuous, which is run by the Dockerfile. The primary motivation for doing this is to allow new commands to be added: primarily a one-manifest mode that will attempt to verify only a single manifest entry. This can be used in our release infra to make sure we are aware of any builds we produce that would not pass this verification.