Command to verify a single manifest
This doesn't perform any of the log operations, but just checks that a manifest is signed correctly, and that the build can be reproduced from the info in the manifest.

Common flags have been moved to the root command to be inherited by all child commands.
This avoids duplication and makes it easier to see what unique flags each command defines.