Allow override of provisioned components
Previously this would only install the latest version of each component from the log. This PR allows the OS to bbe overridden to install a specific version by providing its index in the log. This allows us to test upgrade paths when there are compatibility concerns between the OS and Applet.

If this approach is acceptable, then we can add another flag and extend the GetApplet method to also permit override.

This requires https://github.com/transparency-dev/armored-witness-common/pull/38 to be merged and that dependency updated before this will build.